Refactor out view and alert logic for uncertain delivery
